Possible overflow / divzero bug?
--HG--
branch : beta3_maintenance

+++ src/mechanics/BattleMechanics.h
@@ -43,7 +43,7 @@ class BattleMechanics {
 public:
     virtual bool getCoinFlip(double = 0.5) const = 0;
     virtual unsigned int calculateStat(const Pokemon &p, const STAT i) const = 0;
-    virtual int calculateDamage(BattleField &field, MoveObject &move,
+    virtual unsigned long calculateDamage(BattleField &field, MoveObject &move,
         Pokemon &user, Pokemon &target, const int targets,
         const bool weight = true) const = 0;
     virtual bool attemptHit(BattleField &field, MoveObject &move,
